requirements.txt
dnspython
tqdm
ipaddress
httpx